Fixed many problems with datatype and connection

+++ CaloriesCalculator/CaloriesCalculator.WebClient/Domain/Entities/Product.cs
@@ -4,8 +4,8 @@
     {
         public int Id { get; set; } 
         public string Name { get; set; }
-        public float Proteins { get; set; }
-        public float Fats { get; set; }
-        public float Carbohydrates { get; set; }
+        public decimal Proteins { get; set; }
+        public decimal Fats { get; set; }
+        public decimal Carbohydrates { get; set; }
     }
 }
